refactor: Statistics component for sources
